ASACORE-1526 Fixes to address /W4 compile issues.
Most common warnings to be resolved were:
1. Unreferenced formal parameter. Introduced UNREFERENCED_PARAMETER macros.
2. Assignment operator could not be generated - introduced private assigment operators
3. Conditional expression is constant - loop implemented like 'while(0){}' or 'while(true){}'.
Used for(;;){} or for(;;){break;} instead.
4. Conversion from 'int' to 'unit32_t'('uint8_t', 'DWORD' etc), signed/unsigned mismatch.
Introduced casts.

+++ alljoyn_core/src/windows/KeyStoreListenerFactory.cc
@@ -89,7 +89,7 @@ class DefaultKeyStoreListener : public KeyStoreListener {
                     status = source.PullBytes(dataIn.pbData, dataIn.cbData, pulled);
                 }
 
-                if ((status == ER_OK) && (pulled != fileSize)) {
+                if ((status == ER_OK) && ((int64_t)pulled != fileSize)) {
                     status = ER_BUS_CORRUPT_KEYSTORE;
                 }
